operator

+++ src/Conditions.java
@@ -4,7 +4,7 @@ public class Conditions {
     public static void main(String[] args) {
         Scanner ss = new Scanner(System.in);
         System.out.println("Enter a Number");
-         int num = ss.nextInt();
+         float num = ss.nextFloat();
          if (num<100){
              System.out.println("Noob");
          }else{
